agent/requirements.txt
groq
python-dotenv
requests
streamlit